♻️ refactor: 평균점수, 경쟁률, 학년분포 컴포넌트 분리
