+++ src/components/applicant/application_history/detail/container/RecrutingInfoContainer/Chart/CustomLegend.tsx
@@ -1,4 +1,4 @@
-import { CHARTS_COLORS } from "../../../../../../constants/recruting";
+import { CHARTS_COLORS } from "../../../../../../../constants/recruting";
 
 interface GradeData {
   name: string;
